#98 catch errors produced by invalid timespans
We are currently not validating that both since and until fields of a
timespan are not nil. This shouldn't normally happen, but in ResourceMap
there were some records with created_at = nil. Besides validating this
before storing the timespan, we should be able to deal with already
saved invalid records. This commit performs this check so that they
don't interfere with stats upload.